feat(Collaboration): Main connection gets text from textfield
In our current implementation of the collaboration mode, there is a main connection.
The main connection is the one first connecting to the collaboration server.
This one reads the text from the textarea input field (or another strucutre) and passes it to the tiptap instance.
If only the main connection updates the textdocument other clients connecting later will not overwrite the text.

All clients are syncing changes back to the textarea input.
This also means any client should be able to sumit the content back to the server.

A future addition would be to get/set the text only through the collaboration server which always has the latest state.
If this more sophisticated mehtod is really needed we will implement it later.