fix: Allow to share one toolbar with multiple tiptap instances.
Fix an issue with a toolbar which is shared among multiple editor
instances would invoke the command on each instance.
Now you have to focus at least once the editor so that the toolbar will
issue a command on it.
When clicking a toolbar button, the last focused texteditor instance is
used to issue to command on.
To make sure that the toolbar will always invoke an action even without
first clicking into a tiptap contenteditable area use pat-autofocus.