feat: Late initialization.
Add ``lazy`` argument (default: ``false``) to initialize tiptap only when the text container gets focus.
This is done via an intermediate container which will be replaced when tiptap has finished it's initialization.

This avoids initializing lots of tiptap instances when many text areas are shown, where most of the time none or only a few are likely to be used.
A typical use case is to show a lot of comment boxes with tiptap mentioning functionality.