feat: Introduce tiptap mime type based modes with default extensions. DRAFT /1
NOTE: this would contradict the current mode where only what is defined
in the toolbar is actually allowed as HTML. this is limiting! better, we
define a default set of extension per mime type. this can then also be
imported in the tiptap collaboration server to construct a valid
representation of the content and be able to store that from there back
to the backend.