fix: Fix problem with table not selectable with <ctrl>-<a>
There is a strange problem with tables and selecting all content with
<ctrl>-<a>. When the last node in the document is a table and the last
table cell is empty, then pressing <CTRL>-<a> to select everyting only
selects the very first node in the document. As soon as one of the
conditions - table not the last node or last table cell not empty - is
not met, <CTRL>-<a> selects all, as expected.

We fix the problem by making sure some content is following the table
and add an empty paragraph. When another solution is found we can remove
this hack again.

More information here:
- https://github.com/ueberdosis/tiptap/issues/2401
- https://github.com/ueberdosis/tiptap/issues/3651