fix: Allow to select-all in tables.
Allow to select all (Ctrl-A, Command-A) with tables in certain
conditions.
Due to a tiptap/ProseMirror bug tables with some empty cells and no
content preceding or following the table, selecting the whole table did
not work. This fix makes sure that a non-empty paragraph is added at the
end, if no other content is present at the end of the document.

Follow-up from:
- https://github.com/Patternslib/pat-tiptap/pull/61

More information here:
- https://github.com/ueberdosis/tiptap/issues/2401
- https://github.com/ueberdosis/tiptap/issues/3651