fix: Fix problem with adding columns in tables and deactivate fixed-paragraph.
The fixed-paragraph extension was causing a problem when adding columns.
This extension adds an extra paragraph at the end of tables, so that
it's easier to add content below tables. However, we also have the
"dropcursor" extension activated with tables and other block content for
the same reason - this one adding a horizontal cursor when navigating
below tables which helps in adding content there. To allow adding
columns in tables again, the fixed-paragraph extension is deactivated
until explictly needed again.